split big code source file
