+++ NOTICE
@@ -1,2 +1,2 @@
-AWS Cloud Controller Manager
-Copyright 2018 The Kubernetes Authors.
+OSC Cloud Controller Manager
+Copyright 2020 The 3DS Outscale Authors.
