--- OWNERS
@@ -1,25 +1,0 @@
-approvers:
-- justinsb
-- zmerlynn
-- gnufied
-- jsafrane
-- nckturner
-- micahhausler
-- mcrute
-- gyuho
-- d-nishi
-- M00nF1sh
-- leakingtapan
-reviewers:
-- gnufied
-- jsafrane
-- justinsb
-- zmerlynn
-- chrislovecnm
-- nckturner
-- micahhausler
-- mcrute
-- gyuho
-- d-nishi
-- M00nF1sh
-- leakingtapan
