--- cloud-controller-manager/osc/OWNERS
@@ -1,19 +1,0 @@
-# See the OWNERS docs at https://go.k8s.io/owners
-
-approvers:
-- justinsb
-- zmerlynn
-- gnufied
-- jsafrane
-- micahhausler
-- mcrute
-- m00nf1sh
-reviewers:
-- gnufied
-- jsafrane
-- justinsb
-- zmerlynn
-- chrislovecnm
-- nckturner
-- micahhausler
-- m00nf1sh
