+++ cloud-controller-manager/osc/aws_loadbalancer.go
@@ -16,7 +16,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package aws
+package osc
 
 import (
 	"crypto/sha1"
