+++ cloud-controller-manager/osc/ccm_cloud.go
@@ -1,5 +1,3 @@
-// +build !providerless
-
 /*
 Copyright 2014 The Kubernetes Authors.
 
@@ -16,14 +14,12 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package aws
+package osc
 
 import (
-	"context"
+"context"
 	"errors"
 	"fmt"
-	"io"
-	"net"
 	"path"
 	"sort"
 	"strconv"
@@ -33,17 +29,9 @@ import (
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/awserr"
-	"github.com/aws/aws-sdk-go/aws/credentials"
-	"github.com/aws/aws-sdk-go/aws/ec2metadata"
-	"github.com/aws/aws-sdk-go/aws/endpoints"
-	"github.com/aws/aws-sdk-go/aws/request"
-	"github.com/aws/aws-sdk-go/aws/session"
-	"github.com/aws/aws-sdk-go/service/autoscaling"
 	"github.com/aws/aws-sdk-go/service/ec2"
 	"github.com/aws/aws-sdk-go/service/elb"
 	"github.com/aws/aws-sdk-go/service/elbv2"
-	"github.com/aws/aws-sdk-go/service/kms"
-	"gopkg.in/gcfg.v1"
 	"k8s.io/api/core/v1"
 	"k8s.io/klog"
 
@@ -57,7 +45,6 @@ import (
 	clientset "k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/kubernetes/scheme"
 	v1core "k8s.io/client-go/kubernetes/typed/core/v1"
-	"k8s.io/client-go/pkg/version"
 	"k8s.io/client-go/tools/cache"
 	"k8s.io/client-go/tools/record"
 	"k8s.io/cloud-provider"
@@ -68,1051 +55,7 @@ import (
 	volumehelpers "k8s.io/cloud-provider/volume/helpers"
 )
 
-// ********************* CCM Const Definition *********************
 
-// NLBHealthCheckRuleDescription is the comment used on a security group rule to
-// indicate that it is used for health checks
-const NLBHealthCheckRuleDescription = "kubernetes.io/rule/nlb/health"
-
-// NLBClientRuleDescription is the comment used on a security group rule to
-// indicate that it is used for client traffic
-const NLBClientRuleDescription = "kubernetes.io/rule/nlb/client"
-
-// NLBMtuDiscoveryRuleDescription is the comment used on a security group rule
-// to indicate that it is used for mtu discovery
-const NLBMtuDiscoveryRuleDescription = "kubernetes.io/rule/nlb/mtu"
-
-// ProviderName is the name of this cloud provider.
-const ProviderName = "aws"
-
-// TagNameKubernetesService is the tag name we use to differentiate multiple
-// services. Used currently for ELBs only.
-const TagNameKubernetesService = "kubernetes.io/service-name"
-
-// TagNameSubnetInternalELB is the tag name used on a subnet to designate that
-// it should be used for internal ELBs
-const TagNameSubnetInternalELB = "kubernetes.io/role/internal-elb"
-
-// TagNameSubnetPublicELB is the tag name used on a subnet to designate that
-// it should be used for internet ELBs
-const TagNameSubnetPublicELB = "kubernetes.io/role/elb"
-
-// ServiceAnnotationLoadBalancerType is the annotation used on the service
-// to indicate what type of Load Balancer we want. Right now, the only accepted
-// value is "nlb"
-const ServiceAnnotationLoadBalancerType = "service.beta.kubernetes.io/aws-load-balancer-type"
-
-// ServiceAnnotationLoadBalancerInternal is the annotation used on the service
-// to indicate that we want an internal ELB.
-const ServiceAnnotationLoadBalancerInternal = "service.beta.kubernetes.io/aws-load-balancer-internal"
-
-// ServiceAnnotationLoadBalancerProxyProtocol is the annotation used on the
-// service to enable the proxy protocol on an ELB. Right now we only accept the
-// value "*" which means enable the proxy protocol on all ELB backends. In the
-// future we could adjust this to allow setting the proxy protocol only on
-// certain backends.
-const ServiceAnnotationLoadBalancerProxyProtocol = "service.beta.kubernetes.io/aws-load-balancer-proxy-protocol"
-
-// ServiceAnnotationLoadBalancerAccessLogEmitInterval is the annotation used to
-// specify access log emit interval.
-const ServiceAnnotationLoadBalancerAccessLogEmitInterval = "service.beta.kubernetes.io/aws-load-balancer-access-log-emit-interval"
-
-// ServiceAnnotationLoadBalancerAccessLogEnabled is the annotation used on the
-// service to enable or disable access logs.
-const ServiceAnnotationLoadBalancerAccessLogEnabled = "service.beta.kubernetes.io/aws-load-balancer-access-log-enabled"
-
-// ServiceAnnotationLoadBalancerAccessLogS3BucketName is the annotation used to
-// specify access log s3 bucket name.
-const ServiceAnnotationLoadBalancerAccessLogS3BucketName = "service.beta.kubernetes.io/aws-load-balancer-access-log-s3-bucket-name"
-
-// ServiceAnnotationLoadBalancerAccessLogS3BucketPrefix is the annotation used
-// to specify access log s3 bucket prefix.
-const ServiceAnnotationLoadBalancerAccessLogS3BucketPrefix = "service.beta.kubernetes.io/aws-load-balancer-access-log-s3-bucket-prefix"
-
-// ServiceAnnotationLoadBalancerConnectionDrainingEnabled is the annnotation
-// used on the service to enable or disable connection draining.
-const ServiceAnnotationLoadBalancerConnectionDrainingEnabled = "service.beta.kubernetes.io/aws-load-balancer-connection-draining-enabled"
-
-// ServiceAnnotationLoadBalancerConnectionDrainingTimeout is the annotation
-// used on the service to specify a connection draining timeout.
-const ServiceAnnotationLoadBalancerConnectionDrainingTimeout = "service.beta.kubernetes.io/aws-load-balancer-connection-draining-timeout"
-
-// ServiceAnnotationLoadBalancerConnectionIdleTimeout is the annotation used
-// on the service to specify the idle connection timeout.
-const ServiceAnnotationLoadBalancerConnectionIdleTimeout = "service.beta.kubernetes.io/aws-load-balancer-connection-idle-timeout"
-
-// ServiceAnnotationLoadBalancerCrossZoneLoadBalancingEnabled is the annotation
-// used on the service to enable or disable cross-zone load balancing.
-const ServiceAnnotationLoadBalancerCrossZoneLoadBalancingEnabled = "service.beta.kubernetes.io/aws-load-balancer-cross-zone-load-balancing-enabled"
-
-// ServiceAnnotationLoadBalancerExtraSecurityGroups is the annotation used
-// on the service to specify additional security groups to be added to ELB created
-const ServiceAnnotationLoadBalancerExtraSecurityGroups = "service.beta.kubernetes.io/aws-load-balancer-extra-security-groups"
-
-// ServiceAnnotationLoadBalancerSecurityGroups is the annotation used
-// on the service to specify the security groups to be added to ELB created. Differently from the annotation
-// "service.beta.kubernetes.io/aws-load-balancer-extra-security-groups", this replaces all other security groups previously assigned to the ELB.
-const ServiceAnnotationLoadBalancerSecurityGroups = "service.beta.kubernetes.io/aws-load-balancer-security-groups"
-
-// ServiceAnnotationLoadBalancerCertificate is the annotation used on the
-// service to request a secure listener. Value is a valid certificate ARN.
-// For more, see http://docs.aws.amazon.com/ElasticLoadBalancing/latest/DeveloperGuide/elb-listener-config.html
-// CertARN is an IAM or CM certificate ARN, e.g. arn:aws:acm:us-east-1:123456789012:certificate/12345678-1234-1234-1234-123456789012
-const ServiceAnnotationLoadBalancerCertificate = "service.beta.kubernetes.io/aws-load-balancer-ssl-cert"
-
-// ServiceAnnotationLoadBalancerSSLPorts is the annotation used on the service
-// to specify a comma-separated list of ports that will use SSL/HTTPS
-// listeners. Defaults to '*' (all).
-const ServiceAnnotationLoadBalancerSSLPorts = "service.beta.kubernetes.io/aws-load-balancer-ssl-ports"
-
-// ServiceAnnotationLoadBalancerSSLNegotiationPolicy is the annotation used on
-// the service to specify a SSL negotiation settings for the HTTPS/SSL listeners
-// of your load balancer. Defaults to AWS's default
-const ServiceAnnotationLoadBalancerSSLNegotiationPolicy = "service.beta.kubernetes.io/aws-load-balancer-ssl-negotiation-policy"
-
-// ServiceAnnotationLoadBalancerBEProtocol is the annotation used on the service
-// to specify the protocol spoken by the backend (pod) behind a listener.
-// If `http` (default) or `https`, an HTTPS listener that terminates the
-//  connection and parses headers is created.
-// If set to `ssl` or `tcp`, a "raw" SSL listener is used.
-// If set to `http` and `aws-load-balancer-ssl-cert` is not used then
-// a HTTP listener is used.
-const ServiceAnnotationLoadBalancerBEProtocol = "service.beta.kubernetes.io/aws-load-balancer-backend-protocol"
-
-// ServiceAnnotationLoadBalancerAdditionalTags is the annotation used on the service
-// to specify a comma-separated list of key-value pairs which will be recorded as
-// additional tags in the ELB.
-// For example: "Key1=Val1,Key2=Val2,KeyNoVal1=,KeyNoVal2"
-const ServiceAnnotationLoadBalancerAdditionalTags = "service.beta.kubernetes.io/aws-load-balancer-additional-resource-tags"
-
-// ServiceAnnotationLoadBalancerHCHealthyThreshold is the annotation used on
-// the service to specify the number of successive successful health checks
-// required for a backend to be considered healthy for traffic.
-const ServiceAnnotationLoadBalancerHCHealthyThreshold = "service.beta.kubernetes.io/aws-load-balancer-healthcheck-healthy-threshold"
-
-// ServiceAnnotationLoadBalancerHCUnhealthyThreshold is the annotation used
-// on the service to specify the number of unsuccessful health checks
-// required for a backend to be considered unhealthy for traffic
-const ServiceAnnotationLoadBalancerHCUnhealthyThreshold = "service.beta.kubernetes.io/aws-load-balancer-healthcheck-unhealthy-threshold"
-
-// ServiceAnnotationLoadBalancerHCTimeout is the annotation used on the
-// service to specify, in seconds, how long to wait before marking a health
-// check as failed.
-const ServiceAnnotationLoadBalancerHCTimeout = "service.beta.kubernetes.io/aws-load-balancer-healthcheck-timeout"
-
-// ServiceAnnotationLoadBalancerHCInterval is the annotation used on the
-// service to specify, in seconds, the interval between health checks.
-const ServiceAnnotationLoadBalancerHCInterval = "service.beta.kubernetes.io/aws-load-balancer-healthcheck-interval"
-
-// ServiceAnnotationLoadBalancerEIPAllocations is the annotation used on the
-// service to specify a comma separated list of EIP allocations to use as
-// static IP addresses for the NLB. Only supported on elbv2 (NLB)
-const ServiceAnnotationLoadBalancerEIPAllocations = "service.beta.kubernetes.io/aws-load-balancer-eip-allocations"
-
-// Event key when a volume is stuck on attaching state when being attached to a volume
-const volumeAttachmentStuck = "VolumeAttachmentStuck"
-
-// Indicates that a node has volumes stuck in attaching state and hence it is not fit for scheduling more pods
-const nodeWithImpairedVolumes = "NodeWithImpairedVolumes"
-
-const (
-	// volumeAttachmentConsecutiveErrorLimit is the number of consecutive errors we will ignore when waiting for a volume to attach/detach
-	volumeAttachmentStatusConsecutiveErrorLimit = 10
-	// most attach/detach operations on AWS finish within 1-4 seconds
-	// By using 1 second starting interval with a backoff of 1.8
-	// we get -  [1, 1.8, 3.24, 5.832000000000001, 10.4976]
-	// in total we wait for 2601 seconds
-	volumeAttachmentStatusInitialDelay = 1 * time.Second
-	volumeAttachmentStatusFactor       = 1.8
-	volumeAttachmentStatusSteps        = 13
-
-	// createTag* is configuration of exponential backoff for CreateTag call. We
-	// retry mainly because if we create an object, we cannot tag it until it is
-	// "fully created" (eventual consistency). Starting with 1 second, doubling
-	// it every step and taking 9 steps results in 255 second total waiting
-	// time.
-	createTagInitialDelay = 1 * time.Second
-	createTagFactor       = 2.0
-	createTagSteps        = 9
-
-	// volumeCreate* is configuration of exponential backoff for created volume.
-	// On a random AWS account (shared among several developers) it took 4s on
-	// average, 8s max.
-	volumeCreateInitialDelay  = 5 * time.Second
-	volumeCreateBackoffFactor = 1.2
-	volumeCreateBackoffSteps  = 10
-
-	// Number of node names that can be added to a filter. The AWS limit is 200
-	// but we are using a lower limit on purpose
-	filterNodeLimit = 150
-)
-
-// awsTagNameMasterRoles is a set of well-known AWS tag names that indicate the instance is a master
-// The major consequence is that it is then not considered for AWS zone discovery for dynamic volume creation.
-var awsTagNameMasterRoles = sets.NewString("kubernetes.io/role/master", "k8s.io/role/master")
-
-// Maps from backend protocol to ELB protocol
-var backendProtocolMapping = map[string]string{
-	"https": "https",
-	"http":  "https",
-	"ssl":   "ssl",
-	"tcp":   "ssl",
-}
-
-// MaxReadThenCreateRetries sets the maximum number of attempts we will make when
-// we read to see if something exists and then try to create it if we didn't find it.
-// This can fail once in a consistent system if done in parallel
-// In an eventually consistent system, it could fail unboundedly
-const MaxReadThenCreateRetries = 30
-
-// DefaultVolumeType specifies which storage to use for newly created Volumes
-// TODO: Remove when user/admin can configure volume types and thus we don't
-// need hardcoded defaults.
-const DefaultVolumeType = "gp2"
-
-
-// AWS volume types
-const (
-	// Provisioned IOPS SSD
-	VolumeTypeIO1 = "io1"
-	// General Purpose SSD
-	VolumeTypeGP2 = "gp2"
-	// Cold HDD (sc1)
-	VolumeTypeSC1 = "sc1"
-	// Throughput Optimized HDD
-	VolumeTypeST1 = "st1"
-)
-
-// AWS provisioning limits.
-// Source: http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/EBSVolumeTypes.html
-const (
-	MinTotalIOPS = 100
-	MaxTotalIOPS = 20000
-)
-
-// TagNameClusterNode
-// logically independent clusters running in the same AZ.
-// The tag key = OscK8sNodeName
-// The tag value host name kubernetes.io/hostname
-const TagNameClusterNode = "OscK8sNodeName"
-
-// TagNameClusterNode
-// logically independent clusters running in the same AZ.
-// The tag key = OscK8sNodeName
-// The tag value host name kubernetes.io/hostname
-const TagNameMainSG = "OscK8sMainSG/"
-
-var aliveFilter = []string{
-	ec2.InstanceStateNamePending,
-	ec2.InstanceStateNameRunning,
-	ec2.InstanceStateNameShuttingDown,
-	ec2.InstanceStateNameStopping,
-	ec2.InstanceStateNameStopped,
-}
-
-// ********************* CCM API interfaces *********************
-
-// EC2 is an abstraction over AWS', to allow mocking/other implementations
-// Note that the DescribeX functions return a list, so callers don't need to deal with paging
-// TODO: Should we rename this to AWS (EBS & ELB are not technically part of EC2)
-type EC2 interface {
-	// Query EC2 for instances matching the filter
-	DescribeInstances(request *ec2.DescribeInstancesInput) ([]*ec2.Instance, error)
-
-	// Attach a volume to an instance
-	AttachVolume(*ec2.AttachVolumeInput) (*ec2.VolumeAttachment, error)
-	// Detach a volume from an instance it is attached to
-	DetachVolume(request *ec2.DetachVolumeInput) (resp *ec2.VolumeAttachment, err error)
-	// Lists volumes
-	DescribeVolumes(request *ec2.DescribeVolumesInput) ([]*ec2.Volume, error)
-	// Create an EBS volume
-	CreateVolume(request *ec2.CreateVolumeInput) (resp *ec2.Volume, err error)
-	// Delete an EBS volume
-	DeleteVolume(*ec2.DeleteVolumeInput) (*ec2.DeleteVolumeOutput, error)
-
-	ModifyVolume(*ec2.ModifyVolumeInput) (*ec2.ModifyVolumeOutput, error)
-
-	DescribeVolumeModifications(*ec2.DescribeVolumesModificationsInput) ([]*ec2.VolumeModification, error)
-
-	DescribeSecurityGroups(request *ec2.DescribeSecurityGroupsInput) ([]*ec2.SecurityGroup, error)
-
-	CreateSecurityGroup(*ec2.CreateSecurityGroupInput) (*ec2.CreateSecurityGroupOutput, error)
-	DeleteSecurityGroup(request *ec2.DeleteSecurityGroupInput) (*ec2.DeleteSecurityGroupOutput, error)
-
-	AuthorizeSecurityGroupIngress(*ec2.AuthorizeSecurityGroupIngressInput) (*ec2.AuthorizeSecurityGroupIngressOutput, error)
-	RevokeSecurityGroupIngress(*ec2.RevokeSecurityGroupIngressInput) (*ec2.RevokeSecurityGroupIngressOutput, error)
-
-	DescribeSubnets(*ec2.DescribeSubnetsInput) ([]*ec2.Subnet, error)
-
-	CreateTags(*ec2.CreateTagsInput) (*ec2.CreateTagsOutput, error)
-
-	DescribeRouteTables(request *ec2.DescribeRouteTablesInput) ([]*ec2.RouteTable, error)
-	CreateRoute(request *ec2.CreateRouteInput) (*ec2.CreateRouteOutput, error)
-	DeleteRoute(request *ec2.DeleteRouteInput) (*ec2.DeleteRouteOutput, error)
-
-	ModifyInstanceAttribute(request *ec2.ModifyInstanceAttributeInput) (*ec2.ModifyInstanceAttributeOutput, error)
-
-	DescribeVpcs(input *ec2.DescribeVpcsInput) (*ec2.DescribeVpcsOutput, error)
-}
-
-// ELB is a simple pass-through of AWS' ELB client interface, which allows for testing
-type ELB interface {
-	CreateLoadBalancer(*elb.CreateLoadBalancerInput) (*elb.CreateLoadBalancerOutput, error)
-	DeleteLoadBalancer(*elb.DeleteLoadBalancerInput) (*elb.DeleteLoadBalancerOutput, error)
-	DescribeLoadBalancers(*elb.DescribeLoadBalancersInput) (*elb.DescribeLoadBalancersOutput, error)
-	AddTags(*elb.AddTagsInput) (*elb.AddTagsOutput, error)
-	RegisterInstancesWithLoadBalancer(*elb.RegisterInstancesWithLoadBalancerInput) (*elb.RegisterInstancesWithLoadBalancerOutput, error)
-	DeregisterInstancesFromLoadBalancer(*elb.DeregisterInstancesFromLoadBalancerInput) (*elb.DeregisterInstancesFromLoadBalancerOutput, error)
-	CreateLoadBalancerPolicy(*elb.CreateLoadBalancerPolicyInput) (*elb.CreateLoadBalancerPolicyOutput, error)
-	SetLoadBalancerPoliciesForBackendServer(*elb.SetLoadBalancerPoliciesForBackendServerInput) (*elb.SetLoadBalancerPoliciesForBackendServerOutput, error)
-	SetLoadBalancerPoliciesOfListener(input *elb.SetLoadBalancerPoliciesOfListenerInput) (*elb.SetLoadBalancerPoliciesOfListenerOutput, error)
-	DescribeLoadBalancerPolicies(input *elb.DescribeLoadBalancerPoliciesInput) (*elb.DescribeLoadBalancerPoliciesOutput, error)
-
-	DetachLoadBalancerFromSubnets(*elb.DetachLoadBalancerFromSubnetsInput) (*elb.DetachLoadBalancerFromSubnetsOutput, error)
-	AttachLoadBalancerToSubnets(*elb.AttachLoadBalancerToSubnetsInput) (*elb.AttachLoadBalancerToSubnetsOutput, error)
-
-	CreateLoadBalancerListeners(*elb.CreateLoadBalancerListenersInput) (*elb.CreateLoadBalancerListenersOutput, error)
-	DeleteLoadBalancerListeners(*elb.DeleteLoadBalancerListenersInput) (*elb.DeleteLoadBalancerListenersOutput, error)
-
-	ApplySecurityGroupsToLoadBalancer(*elb.ApplySecurityGroupsToLoadBalancerInput) (*elb.ApplySecurityGroupsToLoadBalancerOutput, error)
-
-	ConfigureHealthCheck(*elb.ConfigureHealthCheckInput) (*elb.ConfigureHealthCheckOutput, error)
-
-	DescribeLoadBalancerAttributes(*elb.DescribeLoadBalancerAttributesInput) (*elb.DescribeLoadBalancerAttributesOutput, error)
-	ModifyLoadBalancerAttributes(*elb.ModifyLoadBalancerAttributesInput) (*elb.ModifyLoadBalancerAttributesOutput, error)
-}
-
-// ELBV2 is a simple pass-through of AWS' ELBV2 client interface, which allows for testing
-type ELBV2 interface {
-	AddTags(input *elbv2.AddTagsInput) (*elbv2.AddTagsOutput, error)
-
-	CreateLoadBalancer(*elbv2.CreateLoadBalancerInput) (*elbv2.CreateLoadBalancerOutput, error)
-	DescribeLoadBalancers(*elbv2.DescribeLoadBalancersInput) (*elbv2.DescribeLoadBalancersOutput, error)
-	DeleteLoadBalancer(*elbv2.DeleteLoadBalancerInput) (*elbv2.DeleteLoadBalancerOutput, error)
-
-	ModifyLoadBalancerAttributes(*elbv2.ModifyLoadBalancerAttributesInput) (*elbv2.ModifyLoadBalancerAttributesOutput, error)
-	DescribeLoadBalancerAttributes(*elbv2.DescribeLoadBalancerAttributesInput) (*elbv2.DescribeLoadBalancerAttributesOutput, error)
-
-	CreateTargetGroup(*elbv2.CreateTargetGroupInput) (*elbv2.CreateTargetGroupOutput, error)
-	DescribeTargetGroups(*elbv2.DescribeTargetGroupsInput) (*elbv2.DescribeTargetGroupsOutput, error)
-	ModifyTargetGroup(*elbv2.ModifyTargetGroupInput) (*elbv2.ModifyTargetGroupOutput, error)
-	DeleteTargetGroup(*elbv2.DeleteTargetGroupInput) (*elbv2.DeleteTargetGroupOutput, error)
-
-	DescribeTargetHealth(input *elbv2.DescribeTargetHealthInput) (*elbv2.DescribeTargetHealthOutput, error)
-
-	DescribeTargetGroupAttributes(*elbv2.DescribeTargetGroupAttributesInput) (*elbv2.DescribeTargetGroupAttributesOutput, error)
-	ModifyTargetGroupAttributes(*elbv2.ModifyTargetGroupAttributesInput) (*elbv2.ModifyTargetGroupAttributesOutput, error)
-
-	RegisterTargets(*elbv2.RegisterTargetsInput) (*elbv2.RegisterTargetsOutput, error)
-	DeregisterTargets(*elbv2.DeregisterTargetsInput) (*elbv2.DeregisterTargetsOutput, error)
-
-	CreateListener(*elbv2.CreateListenerInput) (*elbv2.CreateListenerOutput, error)
-	DescribeListeners(*elbv2.DescribeListenersInput) (*elbv2.DescribeListenersOutput, error)
-	DeleteListener(*elbv2.DeleteListenerInput) (*elbv2.DeleteListenerOutput, error)
-	ModifyListener(*elbv2.ModifyListenerInput) (*elbv2.ModifyListenerOutput, error)
-
-	WaitUntilLoadBalancersDeleted(*elbv2.DescribeLoadBalancersInput) error
-}
-
-// ASG is a simple pass-through of the Autoscaling client interface, which
-// allows for testing.
-type ASG interface {
-	UpdateAutoScalingGroup(*autoscaling.UpdateAutoScalingGroupInput) (*autoscaling.UpdateAutoScalingGroupOutput, error)
-	DescribeAutoScalingGroups(*autoscaling.DescribeAutoScalingGroupsInput) (*autoscaling.DescribeAutoScalingGroupsOutput, error)
-}
-
-// KMS is a simple pass-through of the Key Management Service client interface,
-// which allows for testing.
-type KMS interface {
-	DescribeKey(*kms.DescribeKeyInput) (*kms.DescribeKeyOutput, error)
-}
-
-// EC2Metadata is an abstraction over the AWS metadata service.
-type EC2Metadata interface {
-	Available() bool
-	GetInstanceIdentityDocument() (ec2metadata.EC2InstanceIdentityDocument, error)
-	// Query the EC2 metadata service (used to discover instance-id etc)
-	GetMetadata(path string) (string, error)
-}
-
-// ********************* CCM Used Interfaces *********************
-
-// Services is an abstraction over AWS, to allow mocking/other implementations
-type Services interface {
-	Compute(region string) (EC2, error)
-	LoadBalancing(region string) (ELB, error)
-	LoadBalancingV2(region string) (ELBV2, error)
-	Autoscaling(region string) (ASG, error)
-	Metadata() (EC2Metadata, error)
-	KeyManagement(region string) (KMS, error)
-}
-
-// ********************* CCM Object Interface *********************
-
-// Used to represent a mount device for attaching an EBS volume
-// This should be stored as a single letter (i.e. c, not sdc or /dev/sdc)
-type mountDevice string
-
-// Volumes is an interface for managing cloud-provisioned volumes
-// TODO: Allow other clouds to implement this
-type Volumes interface {
-	// Attach the disk to the node with the specified NodeName
-	// nodeName can be empty to mean "the instance on which we are running"
-	// Returns the device (e.g. /dev/xvdf) where we attached the volume
-	AttachDisk(diskName KubernetesVolumeID, nodeName types.NodeName) (string, error)
-	// Detach the disk from the node with the specified NodeName
-	// nodeName can be empty to mean "the instance on which we are running"
-	// Returns the device where the volume was attached
-	DetachDisk(diskName KubernetesVolumeID, nodeName types.NodeName) (string, error)
-
-	// Create a volume with the specified options
-	CreateDisk(volumeOptions *VolumeOptions) (volumeName KubernetesVolumeID, err error)
-	// Delete the specified volume
-	// Returns true iff the volume was deleted
-	// If the was not found, returns (false, nil)
-	DeleteDisk(volumeName KubernetesVolumeID) (bool, error)
-
-	// Get labels to apply to volume on creation
-	GetVolumeLabels(volumeName KubernetesVolumeID) (map[string]string, error)
-
-	// Get volume's disk path from volume name
-	// return the device path where the volume is attached
-	GetDiskPath(volumeName KubernetesVolumeID) (string, error)
-
-	// Check if the volume is already attached to the node with the specified NodeName
-	DiskIsAttached(diskName KubernetesVolumeID, nodeName types.NodeName) (bool, error)
-
-	// Check if disks specified in argument map are still attached to their respective nodes.
-	DisksAreAttached(map[types.NodeName][]KubernetesVolumeID) (map[types.NodeName]map[KubernetesVolumeID]bool, error)
-
-	// Expand the disk to new size
-	ResizeDisk(diskName KubernetesVolumeID, oldSize resource.Quantity, newSize resource.Quantity) (resource.Quantity, error)
-}
-
-// InstanceGroups is an interface for managing cloud-managed instance groups / autoscaling instance groups
-// TODO: Allow other clouds to implement this
-type InstanceGroups interface {
-	// Set the size to the fixed size
-	ResizeInstanceGroup(instanceGroupName string, size int) error
-	// Queries the cloud provider for information about the specified instance group
-	DescribeInstanceGroup(instanceGroupName string) (InstanceGroupInfo, error)
-}
-
-// InstanceGroupInfo is returned by InstanceGroups.Describe, and exposes information about the group.
-type InstanceGroupInfo interface {
-	// The number of instances currently running under control of this group
-	CurrentSize() (int, error)
-}
-
-// Interface to make the CloudConfig immutable for awsSDKProvider
-type awsCloudConfigProvider interface {
-	getResolver() endpoints.ResolverFunc
-}
-
-// ********************* CCM Object Definitions *********************
-
-// VolumeOptions specifies capacity and tags for a volume.
-type VolumeOptions struct {
-	CapacityGB       int
-	Tags             map[string]string
-	VolumeType       string
-	AvailabilityZone string
-	// IOPSPerGB x CapacityGB will give total IOPS of the volume to create.
-	// Calculated total IOPS will be capped at MaxTotalIOPS.
-	IOPSPerGB int
-	Encrypted bool
-	// fully qualified resource name to the key to use for encryption.
-	// example: arn:aws:kms:us-east-1:012345678910:key/abcd1234-a123-456a-a12b-a123b4cd56ef
-	KmsKeyID string
-}
-
-
-// ********************* CCM Object Init *********************
-
-var _ cloudprovider.Interface = (*Cloud)(nil)
-var _ cloudprovider.Instances = (*Cloud)(nil)
-var _ cloudprovider.LoadBalancer = (*Cloud)(nil)
-var _ cloudprovider.Routes = (*Cloud)(nil)
-var _ cloudprovider.Zones = (*Cloud)(nil)
-var _ cloudprovider.PVLabeler = (*Cloud)(nil)
-
-
-// ********************* CCM Utils functions *********************
-
-// Following functions are used to set outscale endpoints
-func newEC2MetadataSvc() *ec2metadata.EC2Metadata {
-
-	sess := session.Must(session.NewSession(&aws.Config{
-	    EndpointResolver: endpoints.ResolverFunc(OscSetupMetadataResolver()),
-	}))
-	return ec2metadata.New(sess)
-}
-
-func NewMetadata() (MetadataService, error) {
-	svc := newEC2MetadataSvc()
-
-	metadata, err := NewMetadataService(svc)
-	if err != nil {
-		return nil, fmt.Errorf("could not get metadata from AWS: %v", err)
-	}
-	return metadata, err
-}
-
-func NewSession() (*session.Session, error) {
-
-    provider := []credentials.Provider{
-        &credentials.EnvProvider{},
-        &credentials.SharedCredentialsProvider{},
-    }
-
-    metadata,err := NewMetadata()
-    if err != nil {
-        return nil, fmt.Errorf("unable to initialize OSC Metadata session: %v", err)
-    }
-	awsConfig := &aws.Config{
-        Region:                        aws.String(metadata.GetRegion()),
-        Credentials:                   credentials.NewChainCredentials(provider),
-        CredentialsChainVerboseErrors: aws.Bool(true),
-        EndpointResolver: endpoints.ResolverFunc(OscSetupServiceResolver(metadata.GetRegion())),
-    }
-
-    sess, err := session.NewSession(awsConfig)
-    if err != nil {
-        return nil, fmt.Errorf("unable to initialize OSC session: %v", err)
-    }
-    return sess, nil
-}
-
-func newEc2Filter(name string, values ...string) *ec2.Filter {
-	filter := &ec2.Filter{
-		Name: aws.String(name),
-	}
-	for _, value := range values {
-		filter.Values = append(filter.Values, aws.String(value))
-	}
-	return filter
-}
-
-func setNodeDisk(nodeDiskMap map[types.NodeName]map[KubernetesVolumeID]bool, volumeID KubernetesVolumeID,
-				 nodeName types.NodeName, check bool) {
-
-	volumeMap := nodeDiskMap[nodeName]
-
-	if volumeMap == nil {
-		volumeMap = make(map[KubernetesVolumeID]bool)
-		nodeDiskMap[nodeName] = volumeMap
-	}
-	volumeMap[volumeID] = check
-}
-
-// mapNodeNameToPrivateDNSName maps a k8s NodeName to an AWS Instance PrivateDNSName
-// This is a simple string cast
-func mapNodeNameToPrivateDNSName(nodeName types.NodeName) string {
-	return string(nodeName)
-}
-
-// mapInstanceToNodeName maps a EC2 instance to a k8s NodeName, by extracting the PrivateDNSName
-func mapInstanceToNodeName(i *ec2.Instance) types.NodeName {
-	return types.NodeName(aws.StringValue(i.PrivateDnsName))
-}
-
-// Returns the first security group for an instance, or nil
-// We only create instances with one security group, so we don't expect multiple security groups.
-// However, if there are multiple security groups, we will choose the one tagged with our cluster filter.
-// Otherwise we will return an error.
-func findSecurityGroupForInstance(instance *ec2.Instance, taggedSecurityGroups map[string]*ec2.SecurityGroup) (*ec2.GroupIdentifier, error) {
-	instanceID := aws.StringValue(instance.InstanceId)
-
-    klog.Infof("findSecurityGroupForInstance instance.InstanceId : %v", aws.StringValue(instance.InstanceId))
-    klog.Infof("findSecurityGroupForInstance instance.SecurityGroups : %v", instance.SecurityGroups)
-    klog.Infof("findSecurityGroupForInstance taggedSecurityGroups : %v", taggedSecurityGroups)
-
-	var tagged []*ec2.GroupIdentifier
-	var untagged []*ec2.GroupIdentifier
-	for _, group := range instance.SecurityGroups {
-		groupID := aws.StringValue(group.GroupId)
-		if groupID == "" {
-			klog.Warningf("Ignoring security group without id for instance %q: %v", instanceID, group)
-			continue
-		}
-		_, isTagged := taggedSecurityGroups[groupID]
-		if isTagged {
-			tagged = append(tagged, group)
-		} else {
-			untagged = append(untagged, group)
-		}
-	}
-
-	if len(tagged) > 0 {
-		// We create instances with one SG
-		// If users create multiple SGs, they must tag one of them as being k8s owned
-		if len(tagged) != 1 {
-			taggedGroups := ""
-			for _, v := range tagged {
-				taggedGroups += fmt.Sprintf("%s(%s) ", *v.GroupId, *v.GroupName)
-			}
-			return nil, fmt.Errorf("Multiple tagged security groups found for instance %s; ensure only the k8s security group is tagged; the tagged groups were %v", instanceID, taggedGroups)
-		}
-		return tagged[0], nil
-	}
-
-	if len(untagged) > 0 {
-		// For back-compat, we will allow a single untagged SG
-		if len(untagged) != 1 {
-			return nil, fmt.Errorf("Multiple untagged security groups found for instance %s; ensure the k8s security group is tagged", instanceID)
-		}
-		return untagged[0], nil
-	}
-
-	klog.Warningf("No security group found for instance %q", instanceID)
-	return nil, nil
-}
-
-
-// buildListener creates a new listener from the given port, adding an SSL certificate
-// if indicated by the appropriate annotations.
-func buildListener(port v1.ServicePort, annotations map[string]string, sslPorts *portSets) (*elb.Listener, error) {
-	loadBalancerPort := int64(port.Port)
-	portName := strings.ToLower(port.Name)
-	instancePort := int64(port.NodePort)
-	protocol := strings.ToLower(string(port.Protocol))
-	instanceProtocol := protocol
-
-	listener := &elb.Listener{}
-	listener.InstancePort = &instancePort
-	listener.LoadBalancerPort = &loadBalancerPort
-	certID := annotations[ServiceAnnotationLoadBalancerCertificate]
-	if certID != "" && (sslPorts == nil || sslPorts.numbers.Has(loadBalancerPort) || sslPorts.names.Has(portName)) {
-		instanceProtocol = annotations[ServiceAnnotationLoadBalancerBEProtocol]
-		if instanceProtocol == "" {
-			protocol = "ssl"
-			instanceProtocol = "tcp"
-		} else {
-			protocol = backendProtocolMapping[instanceProtocol]
-			if protocol == "" {
-				return nil, fmt.Errorf("Invalid backend protocol %s for %s in %s", instanceProtocol, certID, ServiceAnnotationLoadBalancerBEProtocol)
-			}
-		}
-		listener.SSLCertificateId = &certID
-	} else if annotationProtocol := annotations[ServiceAnnotationLoadBalancerBEProtocol]; annotationProtocol == "http" {
-		instanceProtocol = annotationProtocol
-		protocol = "http"
-	}
-	protocol = strings.ToUpper(protocol)
-	instanceProtocol = strings.ToUpper(instanceProtocol)
-	listener.Protocol = &protocol
-	listener.InstanceProtocol = &instanceProtocol
-
-	return listener, nil
-}
-
-
-func isSubnetPublic(rt []*ec2.RouteTable, subnetID string) (bool, error) {
-	var subnetTable *ec2.RouteTable
-	for _, table := range rt {
-		for _, assoc := range table.Associations {
-			if aws.StringValue(assoc.SubnetId) == subnetID {
-				subnetTable = table
-				break
-			}
-		}
-	}
-
-	if subnetTable == nil {
-		// If there is no explicit association, the subnet will be implicitly
-		// associated with the VPC's main routing table.
-		for _, table := range rt {
-			for _, assoc := range table.Associations {
-				if aws.BoolValue(assoc.Main) == true {
-					klog.V(4).Infof("Assuming implicit use of main routing table %s for %s",
-						aws.StringValue(table.RouteTableId), subnetID)
-					subnetTable = table
-					break
-				}
-			}
-		}
-	}
-
-	if subnetTable == nil {
-		return false, fmt.Errorf("could not locate routing table for subnet %s", subnetID)
-	}
-
-	for _, route := range subnetTable.Routes {
-		// There is no direct way in the AWS API to determine if a subnet is public or private.
-		// A public subnet is one which has an internet gateway route
-		// we look for the gatewayId and make sure it has the prefix of igw to differentiate
-		// from the default in-subnet route which is called "local"
-		// or other virtual gateway (starting with vgv)
-		// or vpc peering connections (starting with pcx).
-		if strings.HasPrefix(aws.StringValue(route.GatewayId), "igw") {
-			return true, nil
-		}
-	}
-
-	return false, nil
-}
-
-type portSets struct {
-	names   sets.String
-	numbers sets.Int64
-}
-
-// getPortSets returns a portSets structure representing port names and numbers
-// that the comma-separated string describes. If the input is empty or equal to
-// "*", a nil pointer is returned.
-func getPortSets(annotation string) (ports *portSets) {
-	if annotation != "" && annotation != "*" {
-		ports = &portSets{
-			sets.NewString(),
-			sets.NewInt64(),
-		}
-		portStringSlice := strings.Split(annotation, ",")
-		for _, item := range portStringSlice {
-			port, err := strconv.Atoi(item)
-			if err != nil {
-				ports.names.Insert(item)
-			} else {
-				ports.numbers.Insert(int64(port))
-			}
-		}
-	}
-	return
-}
-
-func toStatus(lb *elb.LoadBalancerDescription) *v1.LoadBalancerStatus {
-	status := &v1.LoadBalancerStatus{}
-
-	if aws.StringValue(lb.DNSName) != "" {
-		var ingress v1.LoadBalancerIngress
-		ingress.Hostname = aws.StringValue(lb.DNSName)
-		status.Ingress = []v1.LoadBalancerIngress{ingress}
-	}
-
-	return status
-}
-
-func v2toStatus(lb *elbv2.LoadBalancer) *v1.LoadBalancerStatus {
-	status := &v1.LoadBalancerStatus{}
-	if lb == nil {
-		klog.Error("[BUG] v2toStatus got nil input, this is a Kubernetes bug, please report")
-		return status
-	}
-
-	// We check for Active or Provisioning, the only successful statuses
-	if aws.StringValue(lb.DNSName) != "" && (aws.StringValue(lb.State.Code) == elbv2.LoadBalancerStateEnumActive ||
-		aws.StringValue(lb.State.Code) == elbv2.LoadBalancerStateEnumProvisioning) {
-		var ingress v1.LoadBalancerIngress
-		ingress.Hostname = aws.StringValue(lb.DNSName)
-		status.Ingress = []v1.LoadBalancerIngress{ingress}
-	}
-
-	return status
-}
-
-// Finds the value for a given tag.
-func findTag(tags []*ec2.Tag, key string) (string, bool) {
-	for _, tag := range tags {
-		if aws.StringValue(tag.Key) == key {
-			return aws.StringValue(tag.Value), true
-		}
-	}
-	return "", false
-}
-
-
-func isEqualIntPointer(l, r *int64) bool {
-	if l == nil {
-		return r == nil
-	}
-	if r == nil {
-		return l == nil
-	}
-	return *l == *r
-}
-
-func isEqualStringPointer(l, r *string) bool {
-	if l == nil {
-		return r == nil
-	}
-	if r == nil {
-		return l == nil
-	}
-	return *l == *r
-}
-
-func ipPermissionExists(newPermission, existing *ec2.IpPermission, compareGroupUserIDs bool) bool {
-	if !isEqualIntPointer(newPermission.FromPort, existing.FromPort) {
-		return false
-	}
-	if !isEqualIntPointer(newPermission.ToPort, existing.ToPort) {
-		return false
-	}
-	if !isEqualStringPointer(newPermission.IpProtocol, existing.IpProtocol) {
-		return false
-	}
-	// Check only if newPermission is a subset of existing. Usually it has zero or one elements.
-	// Not doing actual CIDR math yet; not clear it's needed, either.
-	klog.V(4).Infof("Comparing %v to %v", newPermission, existing)
-	if len(newPermission.IpRanges) > len(existing.IpRanges) {
-		return false
-	}
-
-	for j := range newPermission.IpRanges {
-		found := false
-		for k := range existing.IpRanges {
-			if isEqualStringPointer(newPermission.IpRanges[j].CidrIp, existing.IpRanges[k].CidrIp) {
-				found = true
-				break
-			}
-		}
-		if !found {
-			return false
-		}
-	}
-
-	for _, leftPair := range newPermission.UserIdGroupPairs {
-		found := false
-		for _, rightPair := range existing.UserIdGroupPairs {
-			if isEqualUserGroupPair(leftPair, rightPair, compareGroupUserIDs) {
-				found = true
-				break
-			}
-		}
-		if !found {
-			return false
-		}
-	}
-
-	return true
-}
-
-func isEqualUserGroupPair(l, r *ec2.UserIdGroupPair, compareGroupUserIDs bool) bool {
-	klog.V(2).Infof("Comparing %v to %v", *l.GroupId, *r.GroupId)
-	if isEqualStringPointer(l.GroupId, r.GroupId) {
-		if compareGroupUserIDs {
-			if isEqualStringPointer(l.UserId, r.UserId) {
-				return true
-			}
-		} else {
-			return true
-		}
-	}
-
-	return false
-}
-
-// Derives the region from a valid az name.
-// Returns an error if the az is known invalid (empty)
-func azToRegion(az string) (string, error) {
-	if len(az) < 1 {
-		return "", fmt.Errorf("invalid (empty) AZ")
-	}
-	region := az[:len(az)-1]
-	return region, nil
-}
-
-
-
-// isRegionValid accepts an AWS region name and returns if the region is a
-// valid region known to the AWS SDK. Considers the region returned from the
-// EC2 metadata service to be a valid region as it's only available on a host
-// running in a valid AWS region.
-func isRegionValid(region string, metadata EC2Metadata) bool {
-	// Does the AWS SDK know about the region?
-	for _, p := range endpoints.DefaultPartitions() {
-		for r := range p.Regions() {
-			if r == region {
-				return true
-			}
-		}
-	}
-
-	// ap-northeast-3 is purposely excluded from the SDK because it
-	// requires an access request (for more details see):
-	// https://github.com/aws/aws-sdk-go/issues/1863
-	if region == "ap-northeast-3" {
-		return true
-	}
-
-	// Fallback to checking if the region matches the instance metadata region
-	// (ignoring any user overrides). This just accounts for running an old
-	// build of Kubernetes in a new region that wasn't compiled into the SDK
-	// when Kubernetes was built.
-	if az, err := getAvailabilityZone(metadata); err == nil {
-		if r, err := azToRegion(az); err == nil && region == r {
-			return true
-		}
-	}
-
-	return false
-}
-
-
-// wrapAttachError wraps the error returned by an AttachVolume request with
-// additional information, if needed and possible.
-func wrapAttachError(err error, disk *awsDisk, instance string) error {
-	if awsError, ok := err.(awserr.Error); ok {
-		if awsError.Code() == "VolumeInUse" {
-			info, err := disk.describeVolume()
-			if err != nil {
-				klog.Errorf("Error describing volume %q: %q", disk.awsID, err)
-			} else {
-				for _, a := range info.Attachments {
-					if disk.awsID != EBSVolumeID(aws.StringValue(a.VolumeId)) {
-						klog.Warningf("Expected to get attachment info of volume %q but instead got info of %q", disk.awsID, aws.StringValue(a.VolumeId))
-					} else if aws.StringValue(a.State) == "attached" {
-						return fmt.Errorf("error attaching EBS volume %q to instance %q: %q. The volume is currently attached to instance %q", disk.awsID, instance, awsError, aws.StringValue(a.InstanceId))
-					}
-				}
-			}
-		}
-	}
-	return fmt.Errorf("error attaching EBS volume %q to instance %q: %q", disk.awsID, instance, err)
-}
-
-func newAWSDisk(aws *Cloud, name KubernetesVolumeID) (*awsDisk, error) {
-	awsID, err := name.MapToAWSVolumeID()
-	if err != nil {
-		return nil, err
-	}
-	disk := &awsDisk{ec2: aws.ec2, name: name, awsID: awsID}
-	return disk, nil
-}
-
-
-// Helper function for describeVolume callers. Tries to retype given error to AWS error
-// and returns true in case the AWS error is "InvalidVolume.NotFound", false otherwise
-func isAWSErrorVolumeNotFound(err error) bool {
-	if err != nil {
-		if awsError, ok := err.(awserr.Error); ok {
-			// https://docs.aws.amazon.com/AWSEC2/latest/APIReference/errors-overview.html
-			if awsError.Code() == "InvalidVolume.NotFound" {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-
-// newAWSInstance creates a new awsInstance object
-func newAWSInstance(ec2Service EC2, instance *ec2.Instance) *awsInstance {
-	az := ""
-	if instance.Placement != nil {
-		az = aws.StringValue(instance.Placement.AvailabilityZone)
-	}
-	self := &awsInstance{
-		ec2:              ec2Service,
-		awsID:            aws.StringValue(instance.InstanceId),
-		nodeName:         mapInstanceToNodeName(instance),
-		availabilityZone: az,
-		instanceType:     aws.StringValue(instance.InstanceType),
-		vpcID:            aws.StringValue(instance.VpcId),
-		subnetID:         aws.StringValue(instance.SubnetId),
-	}
-
-	return self
-}
-
-// extractNodeAddresses maps the instance information from EC2 to an array of NodeAddresses
-func extractNodeAddresses(instance *ec2.Instance) ([]v1.NodeAddress, error) {
-	// Not clear if the order matters here, but we might as well indicate a sensible preference order
-
-	if instance == nil {
-		return nil, fmt.Errorf("nil instance passed to extractNodeAddresses")
-	}
-
-	addresses := []v1.NodeAddress{}
-
-	// handle internal network interfaces
-	for _, networkInterface := range instance.NetworkInterfaces {
-		// skip network interfaces that are not currently in use
-		if aws.StringValue(networkInterface.Status) != ec2.NetworkInterfaceStatusInUse {
-			continue
-		}
-
-		for _, internalIP := range networkInterface.PrivateIpAddresses {
-			if ipAddress := aws.StringValue(internalIP.PrivateIpAddress); ipAddress != "" {
-				ip := net.ParseIP(ipAddress)
-				if ip == nil {
-					return nil, fmt.Errorf("EC2 instance had invalid private address: %s (%q)", aws.StringValue(instance.InstanceId), ipAddress)
-				}
-				addresses = append(addresses, v1.NodeAddress{Type: v1.NodeInternalIP, Address: ip.String()})
-			}
-		}
-	}
-
-	// TODO: Other IP addresses (multiple ips)?
-	publicIPAddress := aws.StringValue(instance.PublicIpAddress)
-	if publicIPAddress != "" {
-		ip := net.ParseIP(publicIPAddress)
-		if ip == nil {
-			return nil, fmt.Errorf("EC2 instance had invalid public address: %s (%s)", aws.StringValue(instance.InstanceId), publicIPAddress)
-		}
-		addresses = append(addresses, v1.NodeAddress{Type: v1.NodeExternalIP, Address: ip.String()})
-	}
-
-	publicDNSName := aws.StringValue(instance.PublicDnsName)
-	if publicDNSName != "" {
-		addresses = append(addresses, v1.NodeAddress{Type: v1.NodeExternalDNS, Address: publicDNSName})
-	}
-
-	return addresses, nil
-}
-
-
-// parseMetadataLocalHostname parses the output of "local-hostname" metadata.
-// If a DHCP option set is configured for a VPC and it has multiple domain names, GetMetadata
-// returns a string containing first the hostname followed by additional domain names,
-// space-separated. For example, if the DHCP option set has:
-// domain-name = us-west-2.compute.internal a.a b.b c.c d.d;
-// $ curl http://169.254.169.254/latest/meta-data/local-hostname
-// ip-192-168-111-51.us-west-2.compute.internal a.a b.b c.c d.d
-func parseMetadataLocalHostname(metadata string) (string, []string) {
-	localHostnames := strings.Fields(metadata)
-	hostname := localHostnames[0]
-	internalDNS := []string{hostname}
-
-	privateAddress := strings.Split(hostname, ".")[0]
-	for _, h := range localHostnames[1:] {
-		internalDNSAddress := privateAddress + "." + h
-		internalDNS = append(internalDNS, internalDNSAddress)
-	}
-	return hostname, internalDNS
-}
-
-func getAvailabilityZone(metadata EC2Metadata) (string, error) {
-	return metadata.GetMetadata("placement/availability-zone")
-}
-
-func updateConfigZone(cfg *CloudConfig, metadata EC2Metadata) error {
-	if cfg.Global.Zone == "" {
-		if metadata != nil {
-			klog.Info("Zone not specified in configuration file; querying AWS metadata service")
-			var err error
-			cfg.Global.Zone, err = getAvailabilityZone(metadata)
-			if err != nil {
-				return err
-			}
-		}
-		if cfg.Global.Zone == "" {
-			return fmt.Errorf("no zone specified in configuration file")
-		}
-	}
-
-	return nil
-}
-
-
-func newAWSSDKProvider(creds *credentials.Credentials, cfg *CloudConfig) *awsSDKProvider {
-	debugPrintCallerFunctionName()
-	klog.V(10).Infof("newAWSSDKProvider(%v,%v)",creds, cfg)
-	return &awsSDKProvider{
-		creds:          creds,
-		cfg:            cfg,
-		regionDelayers: make(map[string]*CrossRequestRetryDelay),
-	}
-}
 
 // ********************* CCM Cloud Object Def *********************
 
@@ -1156,755 +99,6 @@ type Cloud struct {
 	deviceAllocators map[types.NodeName]DeviceAllocator
 }
 
-var _ Volumes = &Cloud{}
-
-
-// ********************* CCM CloudConfig Def & functions *********************
-
-// CloudConfig wraps the settings for the AWS cloud provider.
-// NOTE: Cloud config files should follow the same Kubernetes deprecation policy as
-// flags or CLIs. Config fields should not change behavior in incompatible ways and
-// should be deprecated for at least 2 release prior to removing.
-// See https://kubernetes.io/docs/reference/using-api/deprecation-policy/#deprecating-a-flag-or-cli
-// for more details.
-type CloudConfig struct {
-	Global struct {
-		// TODO: Is there any use for this?  We can get it from the instance metadata service
-		// Maybe if we're not running on AWS, e.g. bootstrap; for now it is not very useful
-		Zone string
-
-		// The AWS VPC flag enables the possibility to run the master components
-		// on a different aws account, on a different cloud provider or on-premises.
-		// If the flag is set also the KubernetesClusterTag must be provided
-		VPC string
-		// SubnetID enables using a specific subnet to use for ELB's
-		SubnetID string
-		// RouteTableID enables using a specific RouteTable
-		RouteTableID string
-
-		// RoleARN is the IAM role to assume when interaction with AWS APIs.
-		RoleARN string
-
-		// KubernetesClusterTag is the legacy cluster id we'll use to identify our cluster resources
-		KubernetesClusterTag string
-		// KubernetesClusterID is the cluster id we'll use to identify our cluster resources
-		KubernetesClusterID string
-
-		//The aws provider creates an inbound rule per load balancer on the node security
-		//group. However, this can run into the AWS security group rule limit of 50 if
-		//many LoadBalancers are created.
-		//
-		//This flag disables the automatic ingress creation. It requires that the user
-		//has setup a rule that allows inbound traffic on kubelet ports from the
-		//local VPC subnet (so load balancers can access it). E.g. 10.82.0.0/16 30000-32000.
-		DisableSecurityGroupIngress bool
-
-		//AWS has a hard limit of 500 security groups. For large clusters creating a security group for each ELB
-		//can cause the max number of security groups to be reached. If this is set instead of creating a new
-		//Security group for each ELB this security group will be used instead.
-		ElbSecurityGroup string
-
-		//During the instantiation of an new AWS cloud provider, the detected region
-		//is validated against a known set of regions.
-		//
-		//In a non-standard, AWS like environment (e.g. Eucalyptus), this check may
-		//be undesirable.  Setting this to true will disable the check and provide
-		//a warning that the check was skipped.  Please note that this is an
-		//experimental feature and work-in-progress for the moment.  If you find
-		//yourself in an non-AWS cloud and open an issue, please indicate that in the
-		//issue body.
-		DisableStrictZoneCheck bool
-	}
-	// [ServiceOverride "1"]
-	//  Service = s3
-	//  Region = region1
-	//  URL = https://s3.foo.bar
-	//  SigningRegion = signing_region
-	//  SigningMethod = signing_method
-	//
-	//  [ServiceOverride "2"]
-	//     Service = ec2
-	//     Region = region2
-	//     URL = https://ec2.foo.bar
-	//     SigningRegion = signing_region
-	//     SigningMethod = signing_method
-	ServiceOverride map[string]*struct {
-		Service       string
-		Region        string
-		URL           string
-		SigningRegion string
-		SigningMethod string
-		SigningName   string
-	}
-}
-
-func (cfg *CloudConfig) validateOverrides() error {
-	if len(cfg.ServiceOverride) == 0 {
-		return nil
-	}
-	set := make(map[string]bool)
-	for onum, ovrd := range cfg.ServiceOverride {
-		// Note: gcfg does not space trim, so we have to when comparing to empty string ""
-		name := strings.TrimSpace(ovrd.Service)
-		if name == "" {
-			return fmt.Errorf("service name is missing [Service is \"\"] in override %s", onum)
-		}
-		// insure the map service name is space trimmed
-		ovrd.Service = name
-
-		region := strings.TrimSpace(ovrd.Region)
-		if region == "" {
-			return fmt.Errorf("service region is missing [Region is \"\"] in override %s", onum)
-		}
-		// insure the map region is space trimmed
-		ovrd.Region = region
-
-		url := strings.TrimSpace(ovrd.URL)
-		if url == "" {
-			return fmt.Errorf("url is missing [URL is \"\"] in override %s", onum)
-		}
-		signingRegion := strings.TrimSpace(ovrd.SigningRegion)
-		if signingRegion == "" {
-			return fmt.Errorf("signingRegion is missing [SigningRegion is \"\"] in override %s", onum)
-		}
-		signature := name + "_" + region
-		if set[signature] {
-			return fmt.Errorf("duplicate entry found for service override [%s] (%s in %s)", onum, name, region)
-		}
-		set[signature] = true
-	}
-	return nil
-}
-
-func (cfg *CloudConfig) getResolver() endpoints.ResolverFunc {
-	defaultResolver := endpoints.DefaultResolver()
-	defaultResolverFn := func(service, region string,
-		optFns ...func(*endpoints.Options)) (endpoints.ResolvedEndpoint, error) {
-		return defaultResolver.EndpointFor(service, region, optFns...)
-	}
-	if len(cfg.ServiceOverride) == 0 {
-		return defaultResolverFn
-	}
-
-	return func(service, region string,
-		optFns ...func(*endpoints.Options)) (endpoints.ResolvedEndpoint, error) {
-		for _, override := range cfg.ServiceOverride {
-			if override.Service == service && override.Region == region {
-				return endpoints.ResolvedEndpoint{
-					URL:           override.URL,
-					SigningRegion: override.SigningRegion,
-					SigningMethod: override.SigningMethod,
-					SigningName:   override.SigningName,
-				}, nil
-			}
-		}
-		return defaultResolver.EndpointFor(service, region, optFns...)
-	}
-}
-
-// ********************* CCM awsSdkEC2 Def & functions *********************
-
-// awsSdkEC2 is an implementation of the EC2 interface, backed by aws-sdk-go
-type awsSdkEC2 struct {
-	ec2 *ec2.EC2
-}
-
-// Implementation of EC2.Instances
-func (s *awsSdkEC2) DescribeInstances(request *ec2.DescribeInstancesInput) ([]*ec2.Instance, error) {
-	// Instances are paged
-	results := []*ec2.Instance{}
-	var nextToken *string
-	requestTime := time.Now()
-	for {
-		response, err := s.ec2.DescribeInstances(request)
-		if err != nil {
-			recordAWSMetric("describe_instance", 0, err)
-			return nil, fmt.Errorf("error listing AWS instances: %q", err)
-		}
-
-		for _, reservation := range response.Reservations {
-			results = append(results, reservation.Instances...)
-		}
-
-		nextToken = response.NextToken
-		if aws.StringValue(nextToken) == "" {
-			break
-		}
-		request.NextToken = nextToken
-	}
-	timeTaken := time.Since(requestTime).Seconds()
-	recordAWSMetric("describe_instance", timeTaken, nil)
-	return results, nil
-}
-
-// Implements EC2.DescribeSecurityGroups
-func (s *awsSdkEC2) DescribeSecurityGroups(request *ec2.DescribeSecurityGroupsInput) ([]*ec2.SecurityGroup, error) {
-	// Security groups are paged
-	results := []*ec2.SecurityGroup{}
-	var nextToken *string
-	requestTime := time.Now()
-	for {
-		response, err := s.ec2.DescribeSecurityGroups(request)
-		if err != nil {
-			recordAWSMetric("describe_security_groups", 0, err)
-			return nil, fmt.Errorf("error listing AWS security groups: %q", err)
-		}
-
-		results = append(results, response.SecurityGroups...)
-
-		nextToken = response.NextToken
-		if aws.StringValue(nextToken) == "" {
-			break
-		}
-		request.NextToken = nextToken
-	}
-	timeTaken := time.Since(requestTime).Seconds()
-	recordAWSMetric("describe_security_groups", timeTaken, nil)
-	return results, nil
-}
-
-func (s *awsSdkEC2) AttachVolume(request *ec2.AttachVolumeInput) (*ec2.VolumeAttachment, error) {
-	requestTime := time.Now()
-	resp, err := s.ec2.AttachVolume(request)
-	timeTaken := time.Since(requestTime).Seconds()
-	recordAWSMetric("attach_volume", timeTaken, err)
-	return resp, err
-}
-
-func (s *awsSdkEC2) DetachVolume(request *ec2.DetachVolumeInput) (*ec2.VolumeAttachment, error) {
-	requestTime := time.Now()
-	resp, err := s.ec2.DetachVolume(request)
-	timeTaken := time.Since(requestTime).Seconds()
-	recordAWSMetric("detach_volume", timeTaken, err)
-	return resp, err
-}
-
-func (s *awsSdkEC2) DescribeVolumes(request *ec2.DescribeVolumesInput) ([]*ec2.Volume, error) {
-	// Volumes are paged
-	results := []*ec2.Volume{}
-	var nextToken *string
-	requestTime := time.Now()
-	for {
-		response, err := s.ec2.DescribeVolumes(request)
-
-		if err != nil {
-			recordAWSMetric("describe_volume", 0, err)
-			return nil, err
-		}
-
-		results = append(results, response.Volumes...)
-
-		nextToken = response.NextToken
-		if aws.StringValue(nextToken) == "" {
-			break
-		}
-		request.NextToken = nextToken
-	}
-	timeTaken := time.Since(requestTime).Seconds()
-	recordAWSMetric("describe_volume", timeTaken, nil)
-	return results, nil
-}
-
-func (s *awsSdkEC2) CreateVolume(request *ec2.CreateVolumeInput) (*ec2.Volume, error) {
-	requestTime := time.Now()
-	resp, err := s.ec2.CreateVolume(request)
-	timeTaken := time.Since(requestTime).Seconds()
-	recordAWSMetric("create_volume", timeTaken, err)
-	return resp, err
-}
-
-func (s *awsSdkEC2) DeleteVolume(request *ec2.DeleteVolumeInput) (*ec2.DeleteVolumeOutput, error) {
-	requestTime := time.Now()
-	resp, err := s.ec2.DeleteVolume(request)
-	timeTaken := time.Since(requestTime).Seconds()
-	recordAWSMetric("delete_volume", timeTaken, err)
-	return resp, err
-}
-
-func (s *awsSdkEC2) ModifyVolume(request *ec2.ModifyVolumeInput) (*ec2.ModifyVolumeOutput, error) {
-	requestTime := time.Now()
-	resp, err := s.ec2.ModifyVolume(request)
-	timeTaken := time.Since(requestTime).Seconds()
-	recordAWSMetric("modify_volume", timeTaken, err)
-	return resp, err
-}
-
-func (s *awsSdkEC2) DescribeVolumeModifications(request *ec2.DescribeVolumesModificationsInput) ([]*ec2.VolumeModification, error) {
-	requestTime := time.Now()
-	results := []*ec2.VolumeModification{}
-	var nextToken *string
-	for {
-		resp, err := s.ec2.DescribeVolumesModifications(request)
-		if err != nil {
-			recordAWSMetric("describe_volume_modification", 0, err)
-			return nil, fmt.Errorf("error listing volume modifictions : %v", err)
-		}
-		results = append(results, resp.VolumesModifications...)
-		nextToken = resp.NextToken
-		if aws.StringValue(nextToken) == "" {
-			break
-		}
-		request.NextToken = nextToken
-	}
-	timeTaken := time.Since(requestTime).Seconds()
-	recordAWSMetric("describe_volume_modification", timeTaken, nil)
-	return results, nil
-}
-
-func (s *awsSdkEC2) DescribeSubnets(request *ec2.DescribeSubnetsInput) ([]*ec2.Subnet, error) {
-	// Subnets are not paged
-	response, err := s.ec2.DescribeSubnets(request)
-	if err != nil {
-		return nil, fmt.Errorf("error listing AWS subnets: %q", err)
-	}
-	return response.Subnets, nil
-}
-
-func (s *awsSdkEC2) CreateSecurityGroup(request *ec2.CreateSecurityGroupInput) (*ec2.CreateSecurityGroupOutput, error) {
-	return s.ec2.CreateSecurityGroup(request)
-}
-
-func (s *awsSdkEC2) DeleteSecurityGroup(request *ec2.DeleteSecurityGroupInput) (*ec2.DeleteSecurityGroupOutput, error) {
-	return s.ec2.DeleteSecurityGroup(request)
-}
-
-func (s *awsSdkEC2) AuthorizeSecurityGroupIngress(request *ec2.AuthorizeSecurityGroupIngressInput) (*ec2.AuthorizeSecurityGroupIngressOutput, error) {
-	return s.ec2.AuthorizeSecurityGroupIngress(request)
-}
-
-func (s *awsSdkEC2) RevokeSecurityGroupIngress(request *ec2.RevokeSecurityGroupIngressInput) (*ec2.RevokeSecurityGroupIngressOutput, error) {
-	return s.ec2.RevokeSecurityGroupIngress(request)
-}
-
-func (s *awsSdkEC2) CreateTags(request *ec2.CreateTagsInput) (*ec2.CreateTagsOutput, error) {
-	debugPrintCallerFunctionName()
-	requestTime := time.Now()
-	resp, err := s.ec2.CreateTags(request)
-	timeTaken := time.Since(requestTime).Seconds()
-	recordAWSMetric("create_tags", timeTaken, err)
-	return resp, err
-}
-
-func (s *awsSdkEC2) DescribeRouteTables(request *ec2.DescribeRouteTablesInput) ([]*ec2.RouteTable, error) {
-	results := []*ec2.RouteTable{}
-	var nextToken *string
-	requestTime := time.Now()
-	for {
-		response, err := s.ec2.DescribeRouteTables(request)
-		if err != nil {
-			recordAWSMetric("describe_route_tables", 0, err)
-			return nil, fmt.Errorf("error listing AWS route tables: %q", err)
-		}
-
-		results = append(results, response.RouteTables...)
-
-		nextToken = response.NextToken
-		if aws.StringValue(nextToken) == "" {
-			break
-		}
-		request.NextToken = nextToken
-	}
-	timeTaken := time.Since(requestTime).Seconds()
-	recordAWSMetric("describe_route_tables", timeTaken, nil)
-	return results, nil
-}
-
-func (s *awsSdkEC2) CreateRoute(request *ec2.CreateRouteInput) (*ec2.CreateRouteOutput, error) {
-	return s.ec2.CreateRoute(request)
-}
-
-func (s *awsSdkEC2) DeleteRoute(request *ec2.DeleteRouteInput) (*ec2.DeleteRouteOutput, error) {
-	return s.ec2.DeleteRoute(request)
-}
-
-func (s *awsSdkEC2) ModifyInstanceAttribute(request *ec2.ModifyInstanceAttributeInput) (*ec2.ModifyInstanceAttributeOutput, error) {
-	return s.ec2.ModifyInstanceAttribute(request)
-}
-
-func (s *awsSdkEC2) DescribeVpcs(request *ec2.DescribeVpcsInput) (*ec2.DescribeVpcsOutput, error) {
-	return s.ec2.DescribeVpcs(request)
-}
-
-// ********************* CCM awsSDKProvider Def & functions *********************
-
-type awsSDKProvider struct {
-	creds *credentials.Credentials
-	cfg   awsCloudConfigProvider
-
-	mutex          sync.Mutex
-	regionDelayers map[string]*CrossRequestRetryDelay
-}
-func (p *awsSDKProvider) addHandlers(regionName string, h *request.Handlers) {
-	h.Build.PushFrontNamed(request.NamedHandler{
-		Name: "k8s/user-agent",
-		Fn:   request.MakeAddToUserAgentHandler("kubernetes", version.Get().String()),
-	})
-
-	h.Sign.PushFrontNamed(request.NamedHandler{
-		Name: "k8s/logger",
-		Fn:   awsHandlerLogger,
-	})
-
-	delayer := p.getCrossRequestRetryDelay(regionName)
-	if delayer != nil {
-		h.Sign.PushFrontNamed(request.NamedHandler{
-			Name: "k8s/delay-presign",
-			Fn:   delayer.BeforeSign,
-		})
-
-		h.AfterRetry.PushFrontNamed(request.NamedHandler{
-			Name: "k8s/delay-afterretry",
-			Fn:   delayer.AfterRetry,
-		})
-	}
-
-	p.addAPILoggingHandlers(h)
-}
-
-func (p *awsSDKProvider) addAPILoggingHandlers(h *request.Handlers) {
-	debugPrintCallerFunctionName()
-	h.Send.PushBackNamed(request.NamedHandler{
-		Name: "k8s/api-request",
-		Fn:   awsSendHandlerLogger,
-	})
-
-	h.ValidateResponse.PushFrontNamed(request.NamedHandler{
-		Name: "k8s/api-validate-response",
-		Fn:   awsValidateResponseHandlerLogger,
-	})
-}
-
-// Get a CrossRequestRetryDelay, scoped to the region, not to the request.
-// This means that when we hit a limit on a call, we will delay _all_ calls to the API.
-// We do this to protect the AWS account from becoming overloaded and effectively locked.
-// We also log when we hit request limits.
-// Note that this delays the current goroutine; this is bad behaviour and will
-// likely cause k8s to become slow or unresponsive for cloud operations.
-// However, this throttle is intended only as a last resort.  When we observe
-// this throttling, we need to address the root cause (e.g. add a delay to a
-// controller retry loop)
-func (p *awsSDKProvider) getCrossRequestRetryDelay(regionName string) *CrossRequestRetryDelay {
-	debugPrintCallerFunctionName()
-	klog.V(10).Infof("getCrossRequestRetryDelay(%v)", regionName)
-	p.mutex.Lock()
-	defer p.mutex.Unlock()
-
-	delayer, found := p.regionDelayers[regionName]
-	if !found {
-		delayer = NewCrossRequestRetryDelay()
-		p.regionDelayers[regionName] = delayer
-	}
-	return delayer
-}
-
-func (p *awsSDKProvider) Compute(regionName string) (EC2, error) {
-	debugPrintCallerFunctionName()
-	klog.V(10).Infof("Compute(%v)", regionName)
-	sess, err := NewSession()
-	if err != nil {
-		return nil, fmt.Errorf("unable to initialize AWS session: %v", err)
-	}
-	service := ec2.New(sess)
-
-	p.addHandlers(regionName, &service.Handlers)
-
-	ec2 := &awsSdkEC2{
-		ec2: service,
-	}
-	return ec2, nil
-}
-
-func (p *awsSDKProvider) LoadBalancing(regionName string) (ELB, error) {
-	debugPrintCallerFunctionName()
-	klog.V(10).Infof("LoadBalancing(%v)", regionName)
-	sess, err := NewSession()
-	if err != nil {
-		return nil, fmt.Errorf("unable to initialize AWS session: %v", err)
-	}
-	elbClient := elb.New(sess)
-	p.addHandlers(regionName, &elbClient.Handlers)
-
-	return elbClient, nil
-}
-
-func (p *awsSDKProvider) LoadBalancingV2(regionName string) (ELBV2, error) {
-	debugPrintCallerFunctionName()
-	klog.V(10).Infof("LoadBalancingV2(%v)", regionName)
-	sess, err := NewSession()
-	if err != nil {
-		return nil, fmt.Errorf("unable to initialize AWS session: %v", err)
-	}
-	elbClient := elbv2.New(sess)
-
-	p.addHandlers(regionName, &elbClient.Handlers)
-
-	return elbClient, nil
-}
-
-func (p *awsSDKProvider) Autoscaling(regionName string) (ASG, error) {
-	debugPrintCallerFunctionName()
-	klog.V(10).Infof("Autoscaling(%v)", regionName)
-	sess, err := NewSession()
-	if err != nil {
-		return nil, fmt.Errorf("unable to initialize AWS session: %v", err)
-	}
-	client := autoscaling.New(sess)
-
-	p.addHandlers(regionName, &client.Handlers)
-
-	return client, nil
-}
-
-func (p *awsSDKProvider) Metadata() (EC2Metadata, error) {
-	debugPrintCallerFunctionName()
-	klog.V(10).Infof("Metadata()")
-	sess, err := session.NewSession(&aws.Config{
-			EndpointResolver: endpoints.ResolverFunc(OscSetupMetadataResolver()),
-    })
-	if err != nil {
-		return nil, fmt.Errorf("unable to initialize OSC session: %v", err)
-	}
-	client := ec2metadata.New(sess)
-	p.addAPILoggingHandlers(&client.Handlers)
-	return client, nil
-}
-
-func (p *awsSDKProvider) KeyManagement(regionName string) (KMS, error) {
-	debugPrintCallerFunctionName()
-	klog.V(10).Infof("KeyManagement(%v)", regionName)
-	sess, err := NewSession()
-	if err != nil {
-		return nil, fmt.Errorf("unable to initialize AWS session: %v", err)
-	}
-	kmsClient := kms.New(sess)
-
-	p.addHandlers(regionName, &kmsClient.Handlers)
-
-	return kmsClient, nil
-}
-
-
-// ********************* CCM awsDisk Def & functions *********************
-
-type awsDisk struct {
-	ec2 EC2
-
-	// Name in k8s
-	name KubernetesVolumeID
-	// id in AWS
-	awsID EBSVolumeID
-}
-
-// Gets the full information about this volume from the EC2 API
-func (d *awsDisk) describeVolume() (*ec2.Volume, error) {
-	volumeID := d.awsID
-
-	request := &ec2.DescribeVolumesInput{
-		VolumeIds: []*string{volumeID.awsString()},
-	}
-
-	volumes, err := d.ec2.DescribeVolumes(request)
-	if err != nil {
-		return nil, err
-	}
-	if len(volumes) == 0 {
-		return nil, fmt.Errorf("no volumes found")
-	}
-	if len(volumes) > 1 {
-		return nil, fmt.Errorf("multiple volumes found")
-	}
-	return volumes[0], nil
-}
-
-func (d *awsDisk) describeVolumeModification() (*ec2.VolumeModification, error) {
-	volumeID := d.awsID
-	request := &ec2.DescribeVolumesModificationsInput{
-		VolumeIds: []*string{volumeID.awsString()},
-	}
-	volumeMods, err := d.ec2.DescribeVolumeModifications(request)
-
-	if err != nil {
-		return nil, fmt.Errorf("error describing volume modification %s with %v", volumeID, err)
-	}
-
-	if len(volumeMods) == 0 {
-		return nil, fmt.Errorf("no volume modifications found for %s", volumeID)
-	}
-	lastIndex := len(volumeMods) - 1
-	return volumeMods[lastIndex], nil
-}
-
-func (d *awsDisk) modifyVolume(requestGiB int64) (int64, error) {
-	volumeID := d.awsID
-
-	request := &ec2.ModifyVolumeInput{
-		VolumeId: volumeID.awsString(),
-		Size:     aws.Int64(requestGiB),
-	}
-	output, err := d.ec2.ModifyVolume(request)
-	if err != nil {
-		modifyError := fmt.Errorf("AWS modifyVolume failed for %s with %v", volumeID, err)
-		return requestGiB, modifyError
-	}
-
-	volumeModification := output.VolumeModification
-
-	if aws.StringValue(volumeModification.ModificationState) == ec2.VolumeModificationStateCompleted {
-		return aws.Int64Value(volumeModification.TargetSize), nil
-	}
-
-	backoff := wait.Backoff{
-		Duration: 1 * time.Second,
-		Factor:   2,
-		Steps:    10,
-	}
-
-	checkForResize := func() (bool, error) {
-		volumeModification, err := d.describeVolumeModification()
-
-		if err != nil {
-			return false, err
-		}
-
-		// According to https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/monitoring_mods.html
-		// Size changes usually take a few seconds to complete and take effect after a volume is in the Optimizing state.
-		if aws.StringValue(volumeModification.ModificationState) == ec2.VolumeModificationStateOptimizing {
-			return true, nil
-		}
-		return false, nil
-	}
-	waitWithErr := wait.ExponentialBackoff(backoff, checkForResize)
-	return requestGiB, waitWithErr
-}
-
-// waitForAttachmentStatus polls until the attachment status is the expected value
-// On success, it returns the last attachment state.
-func (d *awsDisk) waitForAttachmentStatus(status string) (*ec2.VolumeAttachment, error) {
-	backoff := wait.Backoff{
-		Duration: volumeAttachmentStatusInitialDelay,
-		Factor:   volumeAttachmentStatusFactor,
-		Steps:    volumeAttachmentStatusSteps,
-	}
-
-	// Because of rate limiting, we often see errors from describeVolume
-	// So we tolerate a limited number of failures.
-	// But once we see more than 10 errors in a row, we return the error
-	describeErrorCount := 0
-	var attachment *ec2.VolumeAttachment
-
-	err := wait.ExponentialBackoff(backoff, func() (bool, error) {
-		info, err := d.describeVolume()
-		if err != nil {
-			// The VolumeNotFound error is special -- we don't need to wait for it to repeat
-			if isAWSErrorVolumeNotFound(err) {
-				if status == "detached" {
-					// The disk doesn't exist, assume it's detached, log warning and stop waiting
-					klog.Warningf("Waiting for volume %q to be detached but the volume does not exist", d.awsID)
-					stateStr := "detached"
-					attachment = &ec2.VolumeAttachment{
-						State: &stateStr,
-					}
-					return true, nil
-				}
-				if status == "attached" {
-					// The disk doesn't exist, complain, give up waiting and report error
-					klog.Warningf("Waiting for volume %q to be attached but the volume does not exist", d.awsID)
-					return false, err
-				}
-			}
-			describeErrorCount++
-			if describeErrorCount > volumeAttachmentStatusConsecutiveErrorLimit {
-				// report the error
-				return false, err
-			}
-
-			klog.Warningf("Ignoring error from describe volume for volume %q; will retry: %q", d.awsID, err)
-			return false, nil
-		}
-
-		describeErrorCount = 0
-
-		if len(info.Attachments) > 1 {
-			// Shouldn't happen; log so we know if it is
-			klog.Warningf("Found multiple attachments for volume %q: %v", d.awsID, info)
-		}
-		attachmentStatus := ""
-		for _, a := range info.Attachments {
-			if attachmentStatus != "" {
-				// Shouldn't happen; log so we know if it is
-				klog.Warningf("Found multiple attachments for volume %q: %v", d.awsID, info)
-			}
-			if a.State != nil {
-				attachment = a
-				attachmentStatus = *a.State
-			} else {
-				// Shouldn't happen; log so we know if it is
-				klog.Warningf("Ignoring nil attachment state for volume %q: %v", d.awsID, a)
-			}
-		}
-		if attachmentStatus == "" {
-			attachmentStatus = "detached"
-		}
-		if attachmentStatus == status {
-			// Attachment is in requested state, finish waiting
-			return true, nil
-		}
-		// continue waiting
-		klog.V(2).Infof("Waiting for volume %q state: actual=%s, desired=%s", d.awsID, attachmentStatus, status)
-		return false, nil
-	})
-
-	return attachment, err
-}
-
-// Deletes the EBS disk
-func (d *awsDisk) deleteVolume() (bool, error) {
-	request := &ec2.DeleteVolumeInput{VolumeId: d.awsID.awsString()}
-	_, err := d.ec2.DeleteVolume(request)
-	if err != nil {
-		if isAWSErrorVolumeNotFound(err) {
-			return false, nil
-		}
-		if awsError, ok := err.(awserr.Error); ok {
-			if awsError.Code() == "VolumeInUse" {
-				return false, volerr.NewDeletedVolumeInUseError(err.Error())
-			}
-		}
-		return false, fmt.Errorf("error deleting EBS volume %q: %q", d.awsID, err)
-	}
-	return true, nil
-}
-
-// ********************* CCM awsInstance Object & functions *********************
-
-type awsInstance struct {
-	ec2 EC2
-
-	// id in AWS
-	awsID string
-
-	// node name in k8s
-	nodeName types.NodeName
-
-	// availability zone the instance resides in
-	availabilityZone string
-
-	// ID of VPC the instance resides in
-	vpcID string
-
-	// ID of subnet the instance resides in
-	subnetID string
-
-	// instance type
-	instanceType string
-}
-
-// Gets the full information about this instance from the EC2 API
-func (i *awsInstance) describeInstance() (*ec2.Instance, error) {
-	debugPrintCallerFunctionName()
-	klog.V(10).Infof("describeInstance")
-	return describeInstance(i.ec2, InstanceID(i.awsID))
-}
 
 // ********************* CCM Cloud Object functions *********************
 
@@ -4672,175 +2866,3 @@ func (c *Cloud) nodeNameToProviderID(nodeName types.NodeName) (InstanceID, error
 
 	return KubernetesInstanceID(node.Spec.ProviderID).MapToAWSInstanceID()
 }
-
-// ********************* CCM entry point function *********************
-
-// readAWSCloudConfig reads an instance of AWSCloudConfig from config reader.
-func readAWSCloudConfig(config io.Reader) (*CloudConfig, error) {
-	debugPrintCallerFunctionName()
-	klog.V(10).Infof("readAWSCloudConfig(%v,%v)", config)
-	var cfg CloudConfig
-	var err error
-
-	if config != nil {
-		err = gcfg.ReadInto(&cfg, config)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return &cfg, nil
-}
-
-
-// newAWSCloud creates a new instance of AWSCloud.
-// AWSProvider and instanceId are primarily for tests
-func newAWSCloud(cfg CloudConfig, awsServices Services) (*Cloud, error) {
-	debugPrintCallerFunctionName()
-	klog.V(10).Infof("newAWSCloud(%v,%v)", cfg, awsServices)
-	// We have some state in the Cloud object - in particular the attaching map
-	// Log so that if we are building multiple Cloud objects, it is obvious!
-	klog.Infof("Starting OSC cloud provider")
-
-	metadata, err := awsServices.Metadata()
-	if err != nil {
-		return nil, fmt.Errorf("error creating OSC metadata client: %q", err)
-	}
-
-	err = updateConfigZone(&cfg, metadata)
-	if err != nil {
-		return nil, fmt.Errorf("unable to determine OSC zone from cloud provider config or EC2 instance metadata: %v", err)
-	}
-	zone := cfg.Global.Zone
-	if len(zone) <= 1 {
-		return nil, fmt.Errorf("invalid OSC zone in config file: %s", zone)
-	}
-	regionName, err := azToRegion(zone)
-	if err != nil {
-		return nil, err
-	}
-
-	if !cfg.Global.DisableStrictZoneCheck {
-		if !isRegionValid(regionName, metadata) {
-			return nil, fmt.Errorf("not a valid AWS zone (unknown region): %s", zone)
-		}
-	} else {
-		klog.Warningf("Strict OSC zone checking is disabled.  Proceeding with zone: %s", zone)
-	}
-
-	klog.Infof("OSC CCM cfg.Global: %v", cfg.Global)
-	klog.Infof("OSC CCM cfg: %v", cfg)
-
-	klog.Infof("Init Services/Compute")
-	ec2, err := awsServices.Compute(regionName)
-	if err != nil {
-		return nil, fmt.Errorf("error creating OSC EC2 client: %v", err)
-	}
-	klog.Infof("Init Services/LoadBalancing")
-	elb, err := awsServices.LoadBalancing(regionName)
-	if err != nil {
-		return nil, fmt.Errorf("error creating OSC ELB client: %v", err)
-	}
-
-	klog.Infof("Init Services/KeyManagement")
-	kms, err := awsServices.KeyManagement(regionName)
-	if err != nil {
-		return nil, fmt.Errorf("error creating OSC key management client: %v", err)
-	}
-
-	klog.Infof("Init Services/LoadBalancingV2")
-	elbv2, err := awsServices.LoadBalancingV2(regionName)
-	if err != nil {
-		return nil, fmt.Errorf("error creating OSC ELBV2 client: %v", err)
-	}
-	klog.Infof("Init Services/Autoscaling")
-	asg, err := awsServices.Autoscaling(regionName)
-	if err != nil {
-		return nil, fmt.Errorf("error creating AWS autoscaling client: %v", err)
-	}
-
-	awsCloud := &Cloud{
-		ec2:      ec2,
-		elb:      elb,
-		elbv2:    elbv2,
-		asg:      asg,
-		metadata: metadata,
-		kms:      kms,
-		cfg:      &cfg,
-		region:   regionName,
-
-		attaching:        make(map[types.NodeName]map[mountDevice]EBSVolumeID),
-		deviceAllocators: make(map[types.NodeName]DeviceAllocator),
-	}
-	awsCloud.instanceCache.cloud = awsCloud
-
-	tagged := cfg.Global.KubernetesClusterTag != "" || cfg.Global.KubernetesClusterID != ""
-
-	if cfg.Global.VPC != "" && (cfg.Global.SubnetID != "" || cfg.Global.RoleARN != "") && tagged {
-		// When the master is running on a different AWS account, cloud provider or on-premise
-		// build up a dummy instance and use the VPC from the nodes account
-		klog.Info("Master is configured to run on a different AWS account, different cloud provider or on-premises")
-		awsCloud.selfAWSInstance = &awsInstance{
-			nodeName: "master-dummy",
-			vpcID:    cfg.Global.VPC,
-			subnetID: cfg.Global.SubnetID,
-		}
-		awsCloud.vpcID = cfg.Global.VPC
-	} else {
-		selfAWSInstance, err := awsCloud.buildSelfAWSInstance()
-		if err != nil {
-			return nil, err
-		}
-		awsCloud.selfAWSInstance = selfAWSInstance
-		awsCloud.vpcID = selfAWSInstance.vpcID
-		klog.Infof("OSC CCM Instance (%v)", selfAWSInstance)
-		klog.Infof("OSC CCM vpcID (%v)", selfAWSInstance.vpcID)
-
-	}
-
-	if cfg.Global.KubernetesClusterTag != "" || cfg.Global.KubernetesClusterID != "" {
-		if err := awsCloud.tagging.init(cfg.Global.KubernetesClusterTag, cfg.Global.KubernetesClusterID); err != nil {
-			return nil, err
-		}
-	} else {
-		// TODO: Clean up double-API query
-		info, err := awsCloud.selfAWSInstance.describeInstance()
-		if err != nil {
-			return nil, err
-		}
-		if err := awsCloud.tagging.initFromTags(info.Tags); err != nil {
-			return nil, err
-		}
-	}
-	klog.Infof("OSC CCM awsCloud %v", awsCloud)
-	return awsCloud, nil
-}
-
-func init() {
-	debugPrintCallerFunctionName()
-	klog.V(10).Infof("init()")
-	registerMetrics()
-	cloudprovider.RegisterCloudProvider(ProviderName, func(config io.Reader) (cloudprovider.Interface, error) {
-		cfg, err := readAWSCloudConfig(config)
-		if err != nil {
-			return nil, fmt.Errorf("unable to read OSC cloud provider config file: %v", err)
-		}
-
-		if err = cfg.validateOverrides(); err != nil {
-			return nil, fmt.Errorf("unable to validate custom endpoint overrides: %v", err)
-		}
-
-		provider := []credentials.Provider{
-			&credentials.EnvProvider{},
-			&credentials.SharedCredentialsProvider{},
-		}
-
-		creds := credentials.NewChainCredentials(provider)
-
-		aws := newAWSSDKProvider(creds, cfg)
-		return newAWSCloud(*cfg, aws)
-	})
-}
-
-
-
