+++ cloud-controller-manager/osc/doc.go
@@ -14,4 +14,4 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package aws
+package osc
