+++ cloud-controller-manager/osc/sets_ippermissions.go
@@ -16,7 +16,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package aws
+package osc
 
 import (
 	"encoding/json"
