+++ code-of-conduct.md
@@ -1,3 +1,1 @@
-# Kubernetes Community Code of Conduct
 
-Please refer to our [Kubernetes Community Code of Conduct](https://git.k8s.io/community/code-of-conduct.md)
